+++ src/components/ProgressBar/ProgressBar.js
@@ -8,7 +8,8 @@ import { Icon, Loading } from '@folio/stripes/components';
 import { useShowCallout } from '@folio/stripes-acq-components';
 
 import { useBulkOperationDetails } from '../../hooks/api';
-import { EDITING_STEPS, JOB_STATUSES } from '../../constants';
+import { JOB_STATUSES } from '../../constants';
+import { getBulkOperationStep } from './utils';
 
 import css from './ProgressBar.css';
 
@@ -42,24 +43,14 @@ export const ProgressBar = () => {
   };
 
   useEffect(() => {
-    const mappedStatuses = {
-      [JOB_STATUSES.DATA_MODIFICATION]: { step: EDITING_STEPS.UPLOAD },
-      [JOB_STATUSES.COMPLETED]: { step: EDITING_STEPS.COMMIT },
-      [JOB_STATUSES.COMPLETED_WITH_ERRORS]: { step: EDITING_STEPS.UPLOAD },
-      [JOB_STATUSES.FAILED]: { step: null, callout: swwCallout },
-    };
+    const nextStep = getBulkOperationStep(bulkDetails);
 
-    const matched = mappedStatuses[status];
-
-    if (matched) {
-      clearIntervalAndRedirect(`/bulk-edit/${id}/preview`, {
-        ...(matched.step ? { step: matched.step } : {}),
-      });
-
-      if (matched.callout) matched.callout();
+    if (nextStep) {
+      clearIntervalAndRedirect(`/bulk-edit/${id}/preview`, { step: nextStep });
     }
 
     if (status === JOB_STATUSES.FAILED) {
+      swwCallout();
       clearIntervalAndRedirect('/bulk-edit', '');
     }
   }, [status]);