+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Change history for ui-bulk-edit
 
 ## In progress
+
+* [UIBULKED-255](https://issues.folio.org/browse/UIBULKED-255) Incorrect message on the logs landing page.
+* [UIBULKED-265](https://issues.folio.org/browse/UIBULKED-265) Columns on the 'Are you sure' form are mixed in case simultaneously running jobs
+* [UIBULKED-266](https://issues.folio.org/browse/UIBULKED-266) Confirmation screen is missing for completed with errors operations
+* [UIBULKED-252](https://issues.folio.org/browse/UIBULKED-252) Selected filters are not reset on "Logs" tab
+
+## [3.0.2](https://github.com/folio-org/ui-bulk-edit/tree/v3.0.2) (2023-03-02)
+
+* [UIBULKED-253](https://issues.folio.org/browse/UIBULKED-253) User without "Users: Can view user profile" permission can download file from "Logs" tab
+* [UIBULKED-262](https://issues.folio.org/browse/UIBULKED-262) "Clear" action is disabled for Item's permanent location
+
+## [3.0.0](https://github.com/folio-org/ui-bulk-edit/tree/v3.0.0) (2023-02-24)
+
 * [UIBULKED-163](https://issues.folio.org/browse/UIBULKED-163) User can add more rows than supported options
 * [UIBULKED-156](https://issues.folio.org/browse/UIBULKED-156) Add Logs tab to the landing page
 * [UIBULKED-162](https://issues.folio.org/browse/UIBULKED-162) Make In-app Options selection consistent
@@ -26,8 +39,6 @@
 * [UIBULKED-235](https://issues.folio.org/browse/UIBULKED-235) Move @testing-library/dom to dev-deps
 * [UIBULKED-256](https://issues.folio.org/browse/UIBULKED-256) "Expiration date +1 day" is shown on the "Are you sure" form
 * [UIBULKED-257](https://issues.folio.org/browse/UIBULKED-257) Text in the center of "Bulk edit" pane doesn't return to default state when switch from "Identifier" to "Query" tab
-* [UIBULKED-253](https://issues.folio.org/browse/UIBULKED-253) User without "Users: Can view user profile" permission can download file from "Logs" tab
-* [UIBULKED-252](https://issues.folio.org/browse/UIBULKED-252) Selected filters are not reset on "Logs" tab
 
 ## [2.1.3](https://github.com/folio-org/ui-bulk-edit/tree/v2.1.3) (2022-12-02)